Guns now have recoil attribute that can be changed

+++ Assets/Scripts/PlayerOne.cs
@@ -34,7 +34,7 @@ public class PlayerOne : MonoBehaviour
         {
             GameObject weaponmanagerObject = GameObject.Find("PlayerOneWeaponSlot");
             WeaponManager wm = weaponmanagerObject.GetComponent<WeaponManager>();
-            int recoil = wm.Shoot();
+            float recoil = wm.Shoot();
             Recoil(recoil);
 
         }
@@ -43,7 +43,7 @@ public class PlayerOne : MonoBehaviour
     {
         Instantiate(bulletPrefab, firePoint.position, firePoint.rotation);
     }
-    public void Recoil(int recoil)
+    public void Recoil(float recoil)
     {
 
         GetComponent<Rigidbody2D>().AddForce(transform.right * -1 * recoil, ForceMode2D.Impulse);
